+++ app/src/main/java/com/example/mobile_application_course/adapter/StudentViewHolder.kt
@@ -1,6 +1,5 @@
-package com.example.mobile_application_course.recyclerview
+package com.example.mobile_application_course.adapter
 
-import android.util.Log
 import android.view.View
 import android.widget.CheckBox
 import android.widget.TextView
